+++ modules/sui_sdk/src/sui_transaction.cpp
@@ -1,11 +1,11 @@
 #include "sui_transaction.h"
 
-String SuiTransaction::signTransaction(String sender, String recipient, unsigned long long int amount)
+String SuiTransaction::signTransaction(String sender, String recipient, uint64_t amount)
 {
 	return programmable_transaction(sender.utf8().get_data(), recipient.utf8().get_data(), amount);
 }
 
-String SuiTransaction::programmableTransactionAllowSponser(String sender_address, String recipient_address, unsigned long long int amount, String sponser_address)
+String SuiTransaction::programmableTransactionAllowSponser(String sender_address, String recipient_address, uint64_t amount, String sponser_address)
 {
 	return  programmable_transaction_allow_sponser(sender_address.utf8().get_data(), recipient_address.utf8().get_data(), amount, sponser_address.utf8().get_data());
 }
